% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{Trains a list og user specified machine learning models on the data}
\usage{
evaluation(modelSettings, cohortId, outcomeId, outputFolder)
}
\arguments{
\item{cohortId}{The ID of the cohort in the cohort table for which we want to
build covariates.}

\item{outcomeId}{The ID of the outcome cohort in the cohort table for which we want to
predict the occurance of.}

\item{outputFolder}{The directory you want the results to be stored in}

\item{modelList}{A character vector specifying the machine
learning methods to train}
}
\value{
Nothing - the model evaluation is saved in the specified outputFolder
}
\description{
This function evaluates the models across the different datasets and calculates the discriminative performance, model calibration and generalizability of each model.
}
\details{
Given the list of user specified models, the function applies each trained model on the test set for each dataset and
calculates the AUC comparing the prediction with the observered label.  The performance is stored in a csv and ROC plots/calibration plots
are created and saved.
}

