% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractDatasets.R
\name{extractDatasets}
\alias{extractDatasets}
\title{Extracts the training/test data from a list of databases}
\usage{
extractDatasets(connectionDetails, oracleTempSchema = NULL, cdmCohortTable,
  cdmVersion, cdmDatabaseList, cohortId, outcomeId, outputFolder)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{cdmCohortTable}{If not using an existing \code{cohort_person} temp table, what is
the name of the source cohort table?}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and "5".}

\item{cdmDatabaseList}{A list of the database schema that contains the OMOP CDM
you train and evaluate model across.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema, so
for example 'cdm_instance.dbo'.}

\item{cohortId}{The ID of the cohort in the cohort table for which we want to
build covariates.}

\item{outcomeId}{The ID of the outcome cohort in the cohort table for which we want to
predict the occurance of.}

\item{outputFolder}{The directory you want the results to be stored in}
}
\value{
Nothing - the extracted data are saved in the specified outputFolder
}
\description{
User specified cohortId/outcomeId and a list of databases in the OMOP CDM and the code extract the labelled data required to develop a risk prediction model
}
\details{
This function uses the data in the CDM to extract the data consisting of a large set of features for the provided
cohorts. The cohorts are assumed to be in a table with the same structure as the cohort table in
the OMOP CDM. The subject_id in this table must refer to person_ids in the CDM. One person can
occurr multiple times, but the combination of subject_id and cohort_start_date is assumed to be
unique.

This function is called automatically by the \code{\link{extractDatasets}} function.
}

