% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitPredictiveModel2.R
\name{fitPredictiveModel2}
\alias{fitPredictiveModel2}
\title{Trains a list og user specified machine learning models on the data}
\usage{
fitPredictiveModel2(plpData, databaseSchema, outputFolder, modelType, cohortId,
  outcomeId)
}
\arguments{
\item{plpData}{The extracted cohort, outcome and covariate data stored as a list with metadata}

\item{databaseSchema}{The database schema that contains the OMOP CDM to train the model on}

\item{outputFolder}{The directory you want the results to be stored in}

\item{modelType}{The specific machine learning methods to train}

\item{cohortId}{The ID of the cohort in the cohort table for which we want to
build covariates.}

\item{outcomeId}{The ID of the outcome cohort in the cohort table for which we want to
predict the occurance of.}
}
\value{
The train model
}
\description{
This function applies the model training and pre-processing (feature selection)
}
\details{
The function picks the specified model and trains it using the data
}

