% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{gestdiaMain}
\alias{gestdiaMain}
\title{Trains user specified machine learning models on range of user specified datasets and then evaluates the models discriminative performance, calibration and generalizability across all datasets}
\usage{
gestdiaMain(connectionDetails, outputFolder = NULL,
  cdmCohortTable = "cohort", cdmVersion = "5", cohortId = 572,
  outcomeId = 573, cdmDatabaseList = c("cdm_truven_mdcd_v5.dbo"),
  modelList = c("lassLR", "nnet", "randomForest", "gbm"))
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{outputFolder}{The directory you want the results to be stored in}

\item{cdmCohortTable}{If not using an existing \code{cohort_person} temp table, what is
the name of the source cohort table?}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and "5".}

\item{cohortId}{The ID of the cohort in the cohort table for which we want to
build covariates.}

\item{outcomeId}{The ID of the outcome cohort in the cohort table for which we want to
predict the occurance of.}

\item{cdmDatabaseList}{A list of the database schema that contains the OMOP CDM
you train and evaluate model across.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema, so
for example 'cdm_instance.dbo'.}

\item{modelList}{A character vector specifying the machine
                                 learning methods to train}
}
\value{
Nothing - the model evaluation is saved in the specified outputFolder
}
\description{
User specifies cohortId/outcomeId and a list of databases in the OMOP CDM and then the function develops and evaluates numerous risk prediction models to predict the outcome in the cohort population.
}
\details{
This function uses the data in the CDM to develop and evaluate a range of machine leanring techniques for risk prediction.
The cohorts are assumed to be in a table with the same structure as the cohort table in
the OMOP CDM. The subject_id in this table must refer to person_ids in the CDM. One person can
occurr multiple times, but the combination of subject_id and cohort_start_date is assumed to be
unique.
}

