% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions}
\alias{predictions}
\title{Calculates the model predictions on the test set}
\usage{
predictions(cdmDatabaseList, modelList, cohortId, outcomeId, outputFolder,
  quiet = F)
}
\arguments{
\item{modelList}{A character vector specifying the machine
learning methods to train}

\item{cohortId}{The ID of the cohort in the cohort table for which we want to
build covariates.}

\item{outcomeId}{The ID of the outcome cohort in the cohort table for which we want to
predict the occurance of.}

\item{outputFolder}{The directory you want the results to be stored in}

\item{quiet}{Should the function output its progression (TRUE=yes)?}
}
\value{
Nothing - the model evaluation is saved in the specified outputFolder
}
\description{
Calculates the model predictions on the test set
}
\details{
This function calculates the risk predictions on the test set
}

