% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainModels.R
\name{randomForest}
\alias{randomForest}
\title{randomForest: Trains a randomForest model}
\usage{
randomForest(plpdata, databaseSchema, cohortId, outcomeId, outputFolder)
}
\arguments{
\item{databaseSchema}{The database schema that contains the OMOP CDM to train the model on}

\item{cohortId}{The ID of the cohort in the cohort table for which we want to
build covariates.}

\item{outcomeId}{The ID of the outcome cohort in the cohort table for which we want to
predict the occurance of.}

\item{outputFolder}{The directory you want the results to be stored in}

\item{plpData}{The extracted cohort, outcome and covariate data stored as a list with metadata}
}
\value{
Nothing - the trained model and details are saved in the specified outputFolder
}
\description{
This function trains a neural network using the h2o package
}
\details{
...
}

