% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitPredictiveModel2.R
\name{sparseToMat}
\alias{sparseToMat}
\title{Converts the sparse data into a matrix}
\usage{
sparseToMat(plpData, t.databaseSchema, p.databaseSchema, outputFolder)
}
\arguments{
\item{plpData}{The extracted cohort, outcome and covariate data stored as a list with metadata}

\item{t.databaseSchema}{The database schema that contains the OMOP CDM to train the model on}

\item{outputFolder}{The directory you want the results to be stored in}

\item{t.databaseSchema}{The database schema that contains the OMOP CDM to predict the model on}
}
\value{
The file location to the matrix
}
\description{
This function performs wrapper feature selection (using the features choosen by lasso logistic regression)
and then converts the sparse represented data into matrix form
}
\details{
The function performs feature selection on the plpData by finding the feature choosen by the lasso
logistic regression using the training data extracted from the t.databaseSchema and then selecting
all the training data extracted from the p.databaseSchema where the features correspond to the ones
returned by feature selection.  The data are then transformed from sparse format to a matrix with
rows corresponding to people and columns corresponding to features, the matrix is saved in the outputFolder.
}

