% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainModels.R
\name{trainModels}
\alias{trainModels}
\title{Trains a list og user specified machine learning models on the data}
\usage{
trainModels(cdmDatabaseList, modelList, cohortId, outcomeId, outputFolder)
}
\arguments{
\item{modelList}{A character vector specifying the machine
learning methods to train}

\item{cohortId}{The ID of the cohort in the cohort table for which we want to
build covariates.}

\item{outcomeId}{The ID of the outcome cohort in the cohort table for which we want to
predict the occurance of.}

\item{outputFolder}{The directory you want the results to be stored in}
}
\value{
Nothing - the model evaluation is saved in the specified outputFolder
}
\description{
User specifies a list of machine learning models to use and each model is trained automatically using cross validation
}
\details{
This function trains various machine learning models including lasso logistic regression
neural network, random forest, gradient boosting machine and stacking.  Feature selection is applied
for some models to reduce the data dimensionality and improve training speed.
}

